farmtrade_api/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
Pillow==10.1.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
